brisebois/requirements.txt
discord.py[voice]
bs4
requests
pymongo